Development

+++ server/routes/apiroute.ts
@@ -15,7 +15,6 @@ import AdmZip from 'adm-zip';
 import formidable from 'formidable';
 import invaliduploadkey from '../../api/utils/response/invaliduploadkey';
 import { generateRandomString } from '../../utils/generateRandomString';
-import { fileSQL, settingsSQL, shortSQL, userSQL } from '../../api/db/mysql';
 import { isAudio, isImage, isVideo } from '../../utils/mimetypechecker';
 import path from 'path';
 import mime from 'mime-types';
@@ -24,9 +23,14 @@ import dashboardrouter from './dashboard/dashboardroutes';
 import isValidUser from '../../server/middleware/isValidUser';
 import getuploadkey from '../../server/modules/getuploadkey';
 import hasterouter from './haste/hasteroute';
-import { Settings } from '../../types/Dashboard';
 import { validateUploadKey } from '../../api/uploadKey';
+import badrequest from '../../api/utils/response/badrequest';
+import { PrismaClient, Setting } from '@prisma/client';
+import internalservererror from '../../api/utils/response/internalservererror';
+import { permissionsMap } from '../../utils/permissions';
+import missingpermissions from '../../api/utils/response/missingpermissions';
 
+const prisma = new PrismaClient();
 const apirouter = Router();
 
 apirouter.use('/auth', authrouter);
@@ -37,22 +41,52 @@ apirouter.post('/upload', (req: Request, res: Response) => {
    const form = new formidable.IncomingForm();
    form.parse(req, async function (err, fields, files: any) {
       const uploadKey: any = getuploadkey(req) || fields.upload_key;
-      const uploadSetting = await settingsSQL.getSetting('publicUpload');
-      const publicUpload = uploadSetting[0].value === 'true';
+      const publicUploadSetting = await prisma.setting.findUnique({
+         where: {
+            name: 'publicUpload',
+         },
+      });
+
+      if (!publicUploadSetting)
+         return badrequest(res, 'Public upload setting not found');
+
+      const publicUpload = publicUploadSetting.value === 'true';
+
+      let user = null;
 
       if (!publicUpload) {
          if (!(await validateUploadKey(uploadKey as string))) {
             return invaliduploadkey(res);
          }
       }
+
+      if (uploadKey) {
+         user = await prisma.user.findUnique({
+            where: {
+               key: uploadKey,
+            },
+         });
+      }
+      if (user) {
+         const perms = permissionsMap(user.permissions);
+         if (!perms.upload && !publicUpload)
+            return missingpermissions(res, 'upload');
+      }
+
+      if (!files?.file) return badrequest(res, 'No file was uploaded');
+
       const buffer = fs.readFileSync(files?.file?._writeStream?.path);
+      const fileStats = fs.statSync(files?.file?._writeStream?.path);
+      const fileSize = fileStats.size;
+      const fileSizeInMB = fileSize / (1024 * 1024);
+
       fs.unlinkSync(files?.file?._writeStream?.path);
 
       const file = files?.file;
       const newFilename = generateRandomString(15);
       const mimetype = file?.mimetype;
       const originalFilename = file?.originalFilename;
-      const username = uploadKey ? await userSQL.getUser(uploadKey) : 'Unknown';
+      const username = uploadKey && user?.username ? user.username : 'Unknown';
 
       const extension =
          mime.extension(mimetype) || originalFilename.split('.').slice(-1);
@@ -62,89 +96,113 @@ apirouter.post('/upload', (req: Request, res: Response) => {
             paths.image,
             newFilename + '.' + extension,
          );
-         fileSQL.createNewFile(
-            newFilename,
-            mimetype,
-            newFilePath,
-            originalFilename,
-         );
-         uploadFile(
-            paths.image,
+         await prisma.file.create({
+            data: {
+               name: newFilename,
+               mimetype,
+               originalfilename: originalFilename,
+               path: newFilePath,
+               size: fileSize,
+               createdAt: new Date(),
+               ownerId: user ? user.id : null,
+            },
+         });
+         uploadFile({
+            dir: paths.image,
             newFilePath,
             buffer,
-            newFilename,
-            originalFilename,
             mimetype,
-            username,
             extension,
-            'image',
-         );
+            originalFilename,
+            shortname: newFilename,
+            type: 'image',
+            username,
+         });
       } else if (isVideo(mimetype)) {
          const newFilePath = path.resolve(
             paths.video,
             newFilename + '.' + extension,
          );
-         fileSQL.createNewFile(
-            newFilename,
-            mimetype,
-            newFilePath,
-            originalFilename,
-         );
-         uploadFile(
-            paths.video,
+         await prisma.file.create({
+            data: {
+               name: newFilename,
+               mimetype,
+               originalfilename: originalFilename,
+               path: newFilePath,
+               size: fileSize,
+               createdAt: new Date(),
+               ownerId: user ? user.id : null,
+            },
+         });
+         uploadFile({
+            dir: paths.video,
             newFilePath,
             buffer,
-            newFilename,
-            originalFilename,
             mimetype,
-            username,
             extension,
-            'video',
-         );
+            originalFilename,
+            shortname: newFilename,
+            type: 'video',
+            username,
+         });
       } else if (isAudio(mimetype)) {
          const newFilePath = path.resolve(
             paths.audio,
             newFilename + '.' + extension,
          );
-         fileSQL.createNewFile(
-            newFilename,
-            mimetype,
-            newFilePath,
-            originalFilename,
-         );
-         uploadFile(
-            paths.audio,
+         await prisma.file.create({
+            data: {
+               name: newFilename,
+               mimetype,
+               originalfilename: originalFilename,
+               path: newFilePath,
+               size: fileSize,
+               createdAt: new Date(),
+               ownerId: user ? user.id : null,
+            },
+         });
+         uploadFile({
+            dir: paths.audio,
             newFilePath,
             buffer,
-            newFilename,
-            originalFilename,
             mimetype,
-            username,
             extension,
-            'audio',
-         );
+            originalFilename,
+            shortname: newFilename,
+            type: 'video',
+            username,
+         });
       } else {
          const newFilePath = path.resolve(
             paths.data,
             newFilename + '.' + extension,
          );
-         fileSQL.createNewFile(
-            newFilename,
-            'application/zip',
-            newFilePath + '.zip',
-            originalFilename,
-         );
-         uploadFile(
-            paths.data,
+         const shouldZip = fileSizeInMB > 10;
+
+         await prisma.file.create({
+            data: {
+               name: newFilename,
+               mimetype: shouldZip ? 'application/zip' : mimetype,
+               originalfilename: originalFilename,
+               path: shouldZip ? newFilePath + '.zip' : newFilePath,
+               size: fileSize,
+               createdAt: new Date(),
+               ownerId: user ? user.id : null,
+            },
+         });
+
+         uploadFile({
+            dir: paths.data,
             newFilePath,
             buffer,
-            newFilename,
-            originalFilename,
             mimetype,
-            username,
             extension,
-            'data',
-         );
+            originalFilename,
+            shortname: newFilename,
+            type: 'data',
+            username,
+            zip: shouldZip,
+         });
       }
 
       res.statusCode = 201;
@@ -173,13 +231,19 @@ apirouter.post('/upload', (req: Request, res: Response) => {
       fs.writeFileSync(path, buffer);
    }
 
-   function sendNotification(
-      path: string,
-      shortname: string,
-      originalFilename: string,
-      mimetype: string,
-      username: string,
-   ) {
+   function sendNotification({
+      path,
+      shortname,
+      originalFilename,
+      mimetype,
+      username,
+   }: {
+      path: string;
+      shortname: string;
+      originalFilename: string;
+      mimetype: string;
+      username: string;
+   }) {
       if (!sendWebhook) return;
       axiosClient.post(webhooknotification, {
          embeds: [
@@ -199,27 +263,39 @@ apirouter.post('/upload', (req: Request, res: Response) => {
       });
    }
 
-   function uploadFile(
-      dir: string,
-      newFilePath: string,
-      buffer: Buffer,
-      shortname: string,
-      originalFilename: string,
-      mimetype: string,
-      extension: string,
-      username: string,
-      type: 'image' | 'video' | 'audio' | 'data',
-   ): void {
+   function uploadFile({
+      dir,
+      newFilePath,
+      buffer,
+      shortname,
+      originalFilename,
+      mimetype,
+      extension,
+      username,
+      type,
+      zip,
+   }: {
+      dir: string;
+      newFilePath: string;
+      buffer: Buffer;
+      shortname: string;
+      originalFilename: string;
+      mimetype: string;
+      extension: string;
+      username: string;
+      type: 'image' | 'video' | 'audio' | 'data';
+      zip?: boolean;
+   }): void {
       checkIfDirExists(dir);
       saveFile(newFilePath, buffer);
-      sendNotification(
-         newFilePath,
-         shortname,
-         originalFilename,
+      sendNotification({
          mimetype,
+         originalFilename,
+         path: newFilePath,
+         shortname,
          username,
-      );
-      if (type === 'data') {
+      });
+      if (zip) {
          const zip = new AdmZip();
          zip.addLocalFile(newFilePath);
          zip.writeZip(newFilePath + '.zip');
@@ -235,13 +311,33 @@ apirouter.post('/shorter', async (req: Request, res: Response) => {
    const { url } = req.body;
 
    const uploadKey: any = getuploadkey(req);
-   const shorterSetting = await settingsSQL.getSetting('publicShorter');
-   const publicShorter = shorterSetting[0].value === 'true';
+   const shorterSetting = await prisma.setting.findUnique({
+      where: {
+         name: 'publicShorter',
+      },
+   });
+
+   if (!shorterSetting) {
+      return badrequest(res, 'Public shorter setting not found');
+   }
+
+   const publicShorter = shorterSetting.value === 'true';
+
+   let user = null;
 
    if (!publicShorter) {
       if (!(await validateUploadKey(uploadKey as string))) {
          return invaliduploadkey(res);
       }
+
+      user = await prisma.user.findUnique({
+         where: {
+            key: uploadKey,
+         },
+      });
+      if (!user) return badrequest(res);
+      const perms = permissionsMap(user.permissions);
+      if (!perms.shorter) return missingpermissions(res, 'shorter');
    }
 
    if (!validateURL(url)) {
@@ -251,7 +347,15 @@ apirouter.post('/shorter', async (req: Request, res: Response) => {
    }
 
    const short = generateRandomString(6);
-   await shortSQL.createNewShortURL(short, url);
+   await prisma.shorter.create({
+      data: {
+         name: short,
+         url,
+         createdAt: new Date(),
+         ownerId: user ? user.id : null,
+      },
+   });
+
    const shortedURL = `${server}/links/${short}`;
 
    res.statusCode = 200;
@@ -264,13 +368,18 @@ apirouter.get('/upload/:uploadID', async (req: Request, res: Response) => {
 
    if (!uploadID)
       return res.status(400).json({ error: 'No uploadID provided' });
-   const file: any = await fileSQL.selectFile(uploadID as string);
-   if (!file.length) {
+   const file = await prisma.file.findUnique({
+      where: {
+         name: uploadID,
+      },
+   });
+
+   if (!file) {
       return res.status(404).json({ error: 'File not found' });
    } else {
-      const mimetype: string = file[0].mimetype;
-      const filePath: string = file[0].path;
-      const originalFilename: string = file[0].originalfilename;
+      const mimetype: string = file.mimetype;
+      const filePath: string = file.path;
+      const originalFilename: string = file.originalfilename;
 
       if (!mimetype) {
          res.statusMessage = 'No mimetype provided';
@@ -310,12 +419,24 @@ apirouter.get('/links/:link', async (req: Request, res: Response) => {
    const { link } = req.params;
 
    if (!link) return res.status(400).json({ error: 'No link provided' });
-   const shortedURL: any = await shortSQL.selectShortedURL(link as string);
-   if (!shortedURL.length) {
-      return res.status(404).json({ error: 'File not found' });
-   } else {
-      const url = shortedURL[0].url;
+   const shortedURL = await prisma.shorter.findUnique({
+      where: {
+         name: link,
+      },
+   });
 
+   if (!shortedURL) {
+      return res.status(404).json({ error: 'Link not found' });
+   } else {
+      await prisma.shorter.update({
+         where: {
+            name: link,
+         },
+         data: {
+            views: shortedURL.views + 1,
+         },
+      });
+      const url = shortedURL.url;
       res.setHeader('Content-Type', 'application/json');
       return res.status(200).json({
          shortedLink: link,
@@ -325,13 +446,14 @@ apirouter.get('/links/:link', async (req: Request, res: Response) => {
 });
 
 apirouter.get('/settings', async (req: Request, res: Response) => {
-   const settings = await settingsSQL.getAllSettings();
-   const finalSettings: Settings[] = [];
+   const settings = await prisma.setting.findMany();
+   const finalSettings: Setting[] = [];
    settings.forEach((setting) => {
       finalSettings.push({
          name: setting.name,
          value: setting.value,
-         type: setting.type,
+         type: setting.type as any,
+         info: setting.info,
       });
    });
    return res.status(200).json(settings);
